Remove uses of retour-utils in openzt-test-rpc
