+++ Cargo.toml
@@ -17,9 +17,8 @@ members = [
     # "openzt-animations",
 
     # # vendor
-    # "vendor/retour-rs",
-    "vendor/retour-utils",
-    "vendor/retour-utils/impl", "openzt-detour-macro", 
+    "openzt-detour-macro",
+    "openzt-detour", 
 ]
 exclude = ["outside-workspace"]
 resolver = "3"
