+++ openzt-test-rpc/Cargo.toml
@@ -15,6 +15,7 @@ windows =  { version = "0.61.3", features = ["Win32", "Win32_System_Console", "W
 # retour-utils = { path = "../vendor/retour-utils", version = "0.2.1"}
 retour-utils = "0.2.1"
 retour = { version = "0.3.1", features = ["thiscall-abi"] }
+openzt-detour = { path = "../openzt-detour", version = "0.1.0" }
 openzt-detour-macro = { path = "../openzt-detour-macro", version = "0.1.0" }
 
 [lib]
